refactor(cli): separate key bindings config from matching logic
